+++ lib/knossos/algorithm.rb
@@ -1,0 +1,5 @@
+class Knossos::Algorithm
+  def self.carve(grid:)
+    raise NotImplementedError
+  end
+end